requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.5.3

kafka-python==2.0.2

psycopg2-binary==2.9.9

scikit-learn==1.4.0
pandas==2.2.1
numpy==1.26.4

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

loguru==0.7.2
python-dotenv==1.0.1
